General cleanup
The readme was a little out of date and unclear (I found myself not
remembering how to run these / when to run them, so that's a bad sign).

I also removed some cruft from a few of the topic directories which had
either been moved to brew or weren't relevant any longer.