Add python magic
Python is a beast on MacOS but I think I've finally managed to wrangle
it.

This adds a zsh hook to automatically detect the local version in
a given directory and, if that version is not already installed, install
it.  If the version specified is not specific enough for pyenv to be
happy it will install the latest specific version that matches and
update .python-version to use the corrected value.