Automatically configure iTerm
There are so many aspects to iTerm that it's hard to know for sure what
needs configuring, but the good news is that iTerm makes it incredibly
hard to automate modifications to their config outside of the Python
API.

The Python API requires user interaction and also appears to install
a full blown python runtime environment.  Neither of those things fit
the provisioning use case so I've decided to instaed use the "load
preferences from a custom folder or URL" option, which means iTerm is
going to write to the iterm/preferences directory of this repository.

I've included in this commit the R&D work around plist modification that
I did, in case I decide I need to go that route again.  I also explored
use of Dynamic Profiles (which may still be nice to include at some
point).