Don't cycle displays on repeated command
Rectangle had a nifty sounding option where a repeated command on
a window would cycle displays.  The problem is that moving / resizing
/ changing the window wouldn't reset the command -- this means that
simply trying to fix the window BACK to where it was before would be
interpreted as a "repeat" and, in our case, throw it on the other
display.

No good, it is FAR more common for me to want to correct the size of
an adjusted window than to want to move a window to another display.

This fixes it by just disabling the "do something different on repeat"
option entirely.