Replace Dropbox with Maestral
Maestral is a FOSS client for Dropbox, which alone is a good reason to
switch to it.

The other good reason is that the Dropbox client is one of the worst
pieces of software installed on my machine.  It's slow, bloated, spits
ads at me, and took A FULL WEEK TO SYNC.

Wow!  Nowhere to go but up.