Update system defaults
This updates some defaults in ways that work more predictibly, but it
also disables some of the new features that I don't want (apple AI and
window tiling).

It also finally addresses something I've wanted for a while, which is
sane application defaults for various file extensions, and defaulting my
browser to Firefox without me having to spend the 2 seconds clicking
"set as default" the first time I open firefox.